Separate console and structured log exports
