+++ docs/agent-guide.md
@@ -33,22 +33,23 @@ a handler eliminates them.
 
 Use `@briancavalier/fx` for the write-and-run core: `fx`, `Fx`, `Effect`,
 `handle`, `control`, `ok`, `fail`, `tryPromise`, `get`, `run`, `runPromise`,
-`runTask`, and diagnostics helpers. Use named subpaths for optional feature
-areas.
+`runTask`, simple console output, and diagnostics helpers. Use named subpaths
+for optional feature areas.
 
 | Need | Import from |
 | --- | --- |
 | Scopes, abort, finalization, early return, scoped yielding | `@briancavalier/fx/scope` |
 | Structured concurrency | `@briancavalier/fx/concurrent` |
 | Streams and sinks | `@briancavalier/fx/stream` |
-| Time, random, logging | `@briancavalier/fx/time`, `@briancavalier/fx/random`, `@briancavalier/fx/log` |
+| Time and random | `@briancavalier/fx/time`, `@briancavalier/fx/random` |
+| Structured logging | `@briancavalier/fx/log` |
 | Retry and timeout | `@briancavalier/fx/retry`, `@briancavalier/fx/timeout` |
 | HTTP and Node platform boundaries | `@briancavalier/fx/http-client`, `@briancavalier/fx/http-server`, `@briancavalier/fx/platform-node` |
 | Mutable references | `@briancavalier/fx/ref` |
 
-For simple console output, use `consoleLog` with `defaultConsole` from
-`@briancavalier/fx/log`. Use `log`, `info`, `warn`, and `error` from the same
-subpath for structured log messages.
+For simple console output, use `consoleLog` with `defaultConsole` from the root
+import. Use `log`, `info`, `warn`, `error`, and `withConsoleLog` from
+`@briancavalier/fx/log` for structured log messages.
 
 ## Define effects directly
 
@@ -70,8 +71,7 @@ Application logic should request operations and compose results. It should not
 perform platform side effects directly.
 
 ```ts
-import { fx, handle, runPromise } from "@briancavalier/fx"
-import { consoleLog, defaultConsole } from "@briancavalier/fx/log"
+import { consoleLog, defaultConsole, fx, handle, runPromise } from "@briancavalier/fx"
 
 const registerUser = (input: RegisterInput) => fx(function* () {
   const user = validateUser(input)
@@ -137,8 +137,7 @@ Handlers are ordinary pipe transforms. Keep the final interpreter pipeline easy
 to scan.
 
 ```ts
-import { runPromise } from "@briancavalier/fx"
-import { defaultConsole } from "@briancavalier/fx/log"
+import { defaultConsole, runPromise } from "@briancavalier/fx"
 import { defaultTime } from "@briancavalier/fx/time"
 
 const main = program.pipe(
