+++ docs/recipes/run-a-node-program.md
@@ -4,8 +4,8 @@ Use this for Node entrypoints that should run async `Fx` programs and shut down
 cleanly on process signals.
 
 ```ts
+import { defaultConsole } from "@briancavalier/fx"
 import { runNodeMain } from "@briancavalier/fx/platform-node"
-import { defaultConsole } from "@briancavalier/fx/log"
 
 const main = program.pipe(
   defaultConsole
