+++ src/exports.test.ts
@@ -27,12 +27,14 @@ test('root export surface contains small-program core', () => {
   assert.equal(typeof Core.get, 'function')
   assert.equal(typeof Core.provideAll, 'function')
   assert.equal(typeof Core.Task, 'function')
+  assert.equal(typeof Core.consoleLog, 'function')
+  assert.equal(typeof Core.defaultConsole, 'function')
   assert.equal(typeof Core.formatDiagnostic, 'function')
 })
 
 test('root export surface excludes optional feature and platform modules', () => {
   assert.equal('defaultRuntime' in Core, false)
-  assert.equal('defaultConsole' in Core, false)
+  assert.equal('Log' in Core, false)
   assert.equal('defaultRandom' in Core, false)
   assert.equal('defaultTime' in Core, false)
   assert.equal('serve' in Core, false)
@@ -46,9 +48,10 @@ test('feature export surfaces group related functionality', () => {
   assert.equal(typeof Concurrent.all, 'function')
   assert.equal(typeof HttpClient.request, 'function')
   assert.equal(typeof HttpServer.route, 'function')
-  assert.equal(typeof Log.consoleLog, 'function')
-  assert.equal(typeof Log.defaultConsole, 'function')
   assert.equal(typeof Log.log, 'function')
+  assert.equal(typeof Log.withConsoleLog, 'function')
+  assert.equal('consoleLog' in Log, false)
+  assert.equal('defaultConsole' in Log, false)
   assert.equal(typeof PlatformNode.runNodeMain, 'function')
   assert.equal(typeof Random.defaultRandom, 'function')
   assert.equal(typeof Ref.of, 'function')
