+++ src/exports/log.ts
@@ -1,15 +1,6 @@
 export {
-  defaultConsole,
-  error as consoleError,
-  Error as ConsoleError,
-  log as consoleLog,
-  Log as ConsoleLog,
-  type Console
-} from '../Console.js'
-export {
   child,
   collect,
-  console,
   debug,
   error,
   info,
@@ -17,6 +8,7 @@ export {
   log,
   Log,
   minLevel,
+  withConsoleLog,
   warn,
   type LogMessage
 } from '../Log.js'
